Added a test quickjs env, called 'quickjsTest'
This env differs from regular 'quickjs', in that it wraps the computation in the 'runInBoundThread' function.
This is needed because tasty (the testing framework) is run mutithreaded and it seems quickjs does not like it
when it's called from an OS thread other than the one it was started in.
Because Haskell uses lightweight threads, this might happen if threaded mode is enabled, as is the case in testing.
This porblem does not occur when running via Main.hs, since we compile the app as single threaded...

For more info see the paper 'Extending the Haskell Foreign Function Interface with Concurrency'
which can be found at https://simonmar.github.io/bib/papers/conc-ffi.pdf